Use single answer file and model list
Since phased training is using individual model names for judgment, it's hitting the case where multiple models are being loaded for evaluation.  We don't currently need that code path as it was inherited from FastChat so this change removes it and relies on the single answerfile and its model type.

Signed-off-by: Dan McPherson <dmcphers@redhat.com>