Cleanup usage of load model answers
This logic removes the directory walking logic of load_model_answers in favor of finding the exact file needed.  Since we are no longer attempting to support multiple models within a single eval, this simplification is possible and avoids some confusing logic that was in place to allow any judge model to be passed but still use the builtin reference answers.

Signed-off-by: Dan McPherson <dmcphers@redhat.com>