Remove max_workers and serving_gpus from constructor
This change is not backward compatible with callers using max_workers or serving_gpus in the constructor.  A minor version bump should be made for the next release.

Any callers to max_workers= or serving_gpus= in the constructor should switch to use gen_answers and judge_answers for MTBenchEvaluator and MTBenchBranchEvaluator.

Signed-off-by: Dan McPherson <dmcphers@redhat.com>